+++ Src/Object/Item/Installation/MoveVerFloor.h
@@ -46,7 +46,7 @@ private:
 
 	//移動関係
 	float speed_;		//移動速度
-	double distance_;	//ルートまでの距離
+	float distance_;	//ルートまでの距離
 	VECTOR moveVec_;	//移動方向
 
 	//移動
